refactor: Added mapping for status code
