+++ user_service/internal/jsonwebtoken/decode_token.go
@@ -2,7 +2,7 @@ package jsonwebtoken
 
 import (
 	"github.com/golang-jwt/jwt/v5"
-	customerrors "github.com/thestoicway/backend/custom_errors/custom_errors"
+	customerrors "github.com/thestoicway/backend/custom_errors"
 )
 
 // DecodeToken is a method on jwtManager that takes a JWT token as a string,
