+++ user_service/internal/user/database/get_user_by_email.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"errors"
 
-	customerrors "github.com/thestoicway/backend/custom_errors/custom_errors"
+	customerrors "github.com/thestoicway/backend/custom_errors"
 	"github.com/thestoicway/backend/user_service/internal/user/model"
 	"gorm.io/gorm"
 )
